VP-4321: Add additional headers to every AuthHandler (#559)
