+++ VirtoCommerce.Storefront/Infrastructure/Autorest/ClientCredentialsAuthHandler.cs
@@ -5,14 +5,18 @@ using System.Net.Http.Headers;
 using System.Threading;
 using System.Threading.Tasks;
 using IdentityModel.Client;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Options;
+using VirtoCommerce.Storefront.Model;
 
 namespace VirtoCommerce.Storefront.Infrastructure.Autorest
 {
     /// <summary>
-    /// HTTP message delegating handler that encapsulates access token handling and renewal
+    /// HTTP message delegating handler that encapsulates access token handling and renewment
+    /// Implements id-secret authorization to the Platform API 
     /// </summary>
-    public class ClientCredentialsAuthHandler : DelegatingHandler
+
+    public class ClientCredentialsAuthHandler : BaseAuthHandler
     {
         private static readonly SemaphoreSlim _lock = new SemaphoreSlim(1, 1);
         private static string _accessToken;
@@ -53,7 +57,7 @@ namespace VirtoCommerce.Storefront.Infrastructure.Autorest
         /// </summary>
         /// <param name="options"></param>
         /// <param name="clientFactory"></param>
-        public ClientCredentialsAuthHandler(IOptions<PlatformEndpointOptions> options, IHttpClientFactory clientFactory)
+        public ClientCredentialsAuthHandler(IOptions<PlatformEndpointOptions> options, IHttpClientFactory clientFactory, IWorkContextAccessor workContextAccessor, IHttpContextAccessor httpContextAccessor) : base(workContextAccessor, httpContextAccessor)
         {
             _options = options.Value;
             _clientFactory = clientFactory;
@@ -75,7 +79,6 @@ namespace VirtoCommerce.Storefront.Infrastructure.Autorest
                 return new HttpResponseMessage(HttpStatusCode.Unauthorized) { RequestMessage = request };
             }
 
-            request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", AccessToken);
             var response = await base.SendAsync(request, cancellationToken).ConfigureAwait(false);
 
             if (response.StatusCode != HttpStatusCode.Unauthorized)
@@ -90,8 +93,12 @@ namespace VirtoCommerce.Storefront.Infrastructure.Autorest
 
             response.Dispose(); // This 401 response will not be used for anything so is disposed to unblock the socket.
 
-            request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", AccessToken);
             return await base.SendAsync(request, cancellationToken).ConfigureAwait(false);
+        }
+
+        protected override void AddAuthentication(HttpRequestMessage request)
+        {
+            request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", AccessToken);
         }
 
         private async Task<bool> RenewTokensAsync(CancellationToken cancellationToken)
